+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## master
 
+* `core` – Raise `Shrine::NotCached` in `Attacher#attach_cached` when non-cached uploaded file is assigned (@janko)
+
+* `allow_reassign` – Add new plugin for ignoring exception when assigning stored file if it's the current file (@janko)
+
 * `derivatives` – Fix `:derivative` value inconsistency when derivatives are being promoted (@janko)
 
 * `add_metadata` – Add `#add_metadata` method for adding metadata to uploaded files (@janko)